+++ common/int64.hh
@@ -91,7 +91,7 @@ namespace csl
         inline const unsigned char * ucharp_data() const { return reinterpret_cast<const unsigned char *>(&value_); }
 
         /** @brief returns the size of the variable data */
-        inline uint64_t var_size() const { return sizeof(value_); }
+        inline size_t var_size() const { return sizeof(value_); }
 
         /* conversions to other types */
         /**
@@ -177,7 +177,7 @@ namespace csl
         this function assumes that v has enough space. sz will be updated to indicate, how
         many bytes were stored.
          */
-        bool to_binary(unsigned char * v, uint64_t & sz) const;
+        bool to_binary(unsigned char * v, size_t & sz) const;
 
         /**
         @brief convert to (ptr+size)
@@ -188,7 +188,7 @@ namespace csl
         this function assumes that v has enough space. sz will be updated to indicate, how
         many bytes were stored.
          */
-        bool to_binary(void * v, uint64_t & sz) const;
+        bool to_binary(void * v, size_t & sz) const;
 
         /**
         @brief stores the content of this instance to an XDR stream
@@ -306,7 +306,7 @@ namespace csl
 
         this function delegates the conversion task to the common::binry class
          */
-        bool from_binary(const unsigned char * v,uint64_t sz);
+        bool from_binary(const unsigned char * v,size_t sz);
 
         /**
         @brief convert a memory region
@@ -315,7 +315,7 @@ namespace csl
 
         this function delegates the conversion task to the common::binry class
          */
-        bool from_binary(const void * v,uint64_t sz);
+        bool from_binary(const void * v,size_t sz);
 
         /**
         @brief read the content of this instance from an XDR stream
